restructure of controller package structure
as there are limitations given regarding the grouping of parameters for swagger docs
