+++ src/main/lombok/org/heigit/ohsome/ohsomeapi/controller/dataaggregation/elements/PerimeterController.java
@@ -1,4 +1,4 @@
-package org.heigit.ohsome.ohsomeapi.controller.elements.dataaggregation;
+package org.heigit.ohsome.ohsomeapi.controller.dataaggregation.elements;
 
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiImplicitParam;
@@ -20,17 +20,18 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
+
 /**
- * Controller containing the GET and POST request handling methods, which are mapped to
- * "/elements/length".
+ * Controller containing the GET and POST servletRequest handling methods, which are mapped to
+ * "/elements/perimeter".
  */
-@Api(tags = "Length")
+@Api(tags = "Perimeter")
 @RestController
-@RequestMapping("/elements/length")
-public class LengthController {
+@RequestMapping("/elements/perimeter")
+public class PerimeterController {
 
   /**
-   * Gives the length of OSM elements.
+   * Gives the perimeter of polygonal OSM objects.
    * 
    * @param servletRequest <code>HttpServletRequest</code> of the incoming request
    * @param servletResponse <code>HttpServletResponse</code> of the outgoing response
@@ -38,19 +39,19 @@ public class LengthController {
    * @throws Exception thrown by
    *         {@link org.heigit.ohsome.ohsomeapi.executor.AggregateRequestExecutor#aggregate() aggregate}
    */
-  @ApiOperation(value = "Length of OSM elements", nickname = "length",
+  @ApiOperation(value = "Perimeter of OSM elements", nickname = "perimeter",
       response = DefaultAggregationResponse.class)
   @RequestMapping(value = "", method = {RequestMethod.GET, RequestMethod.POST},
       produces = {"application/json", "text/csv"})
-  public Response length(HttpServletRequest servletRequest, HttpServletResponse servletResponse)
+  public Response perimeter(HttpServletRequest servletRequest, HttpServletResponse servletResponse)
       throws Exception {
-    AggregateRequestExecutor executor =
-        new AggregateRequestExecutor(RequestResource.LENGTH, servletRequest, servletResponse, false);
+    AggregateRequestExecutor executor = new AggregateRequestExecutor(RequestResource.PERIMETER,
+        servletRequest, servletResponse, false);
     return executor.aggregate();
   }
 
   /**
-   * Gives the length of OSM elements grouped by the OSM type.
+   * Gives the perimeter of polygonal OSM objects grouped by the OSM type.
    * 
    * @param servletRequest <code>HttpServletRequest</code> of the incoming request
    * @param servletResponse <code>HttpServletResponse</code> of the outgoing response
@@ -59,18 +60,21 @@ public class LengthController {
    *         {@link org.heigit.ohsome.ohsomeapi.executor.ElementsRequestExecutor#aggregateGroupByType(RequestResource, HttpServletRequest, HttpServletResponse, boolean, boolean)
    *         aggregateGroupByType}
    */
-  @ApiOperation(value = "Length of OSM elements grouped by the type",
-      nickname = "lengthGroupByType", response = GroupByResponse.class)
+  @ApiOperation(value = "Perimeter of OSM elements grouped by the type",
+      nickname = "perimeterGroupByType", response = GroupByResponse.class)
+  @ApiImplicitParam(name = "filter", value = ParameterDescriptions.FILTER,
+      defaultValue = DefaultSwaggerParameters.BUILDING_FILTER, paramType = "query",
+      dataType = "string", required = false)
   @RequestMapping(value = "/groupBy/type", method = {RequestMethod.GET, RequestMethod.POST},
       produces = {"application/json", "text/csv"})
-  public Response lengthGroupByType(HttpServletRequest servletRequest,
+  public Response perimeterGroupByType(HttpServletRequest servletRequest,
       HttpServletResponse servletResponse) throws Exception {
-    return ElementsRequestExecutor.aggregateGroupByType(RequestResource.LENGTH, servletRequest,
+    return ElementsRequestExecutor.aggregateGroupByType(RequestResource.PERIMETER, servletRequest,
         servletResponse, true, false);
   }
 
   /**
-   * Gives the length of OSM elements grouped by the boundary parameter (bounding
+   * Gives the perimeter of polygonal OSM objects grouped by the boundary parameter (bounding
    * box/circle/polygon).
    * 
    * @param servletRequest <code>HttpServletRequest</code> of the incoming request
@@ -80,21 +84,20 @@ public class LengthController {
    *         {@link org.heigit.ohsome.ohsomeapi.executor.AggregateRequestExecutor#aggregateGroupByBoundary() aggregateGroupByBoundary}
    */
   @ApiOperation(
-      value = "Length of OSM elements grouped by the boundary (bboxes, bcircles, or bpolys)",
-      nickname = "lengthGroupByBoundary", response = GroupByResponse.class)
-  @ApiImplicitParam(name = "format", value = ParameterDescriptions.FORMAT, defaultValue = "",
-      paramType = "query", dataType = "string", required = false)
+      value = "Perimeter of OSM elements in grouped by the boundary "
+          + "(bboxes, bcircles, or bpolys)",
+      nickname = "perimeterGroupByBoundary", response = GroupByResponse.class)
   @RequestMapping(value = "/groupBy/boundary", method = {RequestMethod.GET, RequestMethod.POST},
       produces = {"application/json", "text/csv"})
-  public Response lengthGroupByBoundary(HttpServletRequest servletRequest,
+  public Response perimeterGroupByBoundary(HttpServletRequest servletRequest,
       HttpServletResponse servletResponse) throws Exception {
-    AggregateRequestExecutor executor =
-        new AggregateRequestExecutor(RequestResource.LENGTH, servletRequest, servletResponse, false);
+    AggregateRequestExecutor executor = new AggregateRequestExecutor(RequestResource.PERIMETER,
+        servletRequest, servletResponse, false);
     return executor.aggregateGroupByBoundary();
   }
 
   /**
-   * Gives the length of OSM elements grouped by the boundary and the tag.
+   * Gives the perimeter of polygonal OSM objects grouped by the boundary and the tag.
    * 
    * @param servletRequest <code>HttpServletRequest</code> of the incoming request
    * @param servletResponse <code>HttpServletResponse</code> of the outgoing response
@@ -103,8 +106,8 @@ public class LengthController {
    *         {@link org.heigit.ohsome.ohsomeapi.executor.ElementsRequestExecutor#aggregateGroupByBoundaryGroupByTag(RequestResource, HttpServletRequest, HttpServletResponse, boolean, boolean)
    *         aggregateGroupByBoundaryGroupByTag}
    */
-  @ApiOperation(value = "Length of OSM elements grouped by the boundary and the tag",
-      nickname = "lengthGroupByBoundaryGroupByTag", response = GroupByResponse.class)
+  @ApiOperation(value = "Perimeter of OSM elements grouped by the boundary and the tag",
+      nickname = "perimeterGroupByBoundaryGroupByTag", response = GroupByResponse.class)
   @ApiImplicitParams({
       @ApiImplicitParam(name = "groupByKey", value = ParameterDescriptions.GROUP_BY_KEY,
           defaultValue = DefaultSwaggerParameters.BUILDING_KEY, paramType = "query",
@@ -113,14 +116,14 @@ public class LengthController {
           defaultValue = "", paramType = "query", dataType = "string", required = false)})
   @RequestMapping(value = "/groupBy/boundary/groupBy/tag",
       method = {RequestMethod.GET, RequestMethod.POST}, produces = {"application/json", "text/csv"})
-  public Response lengthGroupByBoundaryGroupByTag(HttpServletRequest servletRequest,
+  public Response perimeterGroupByBoundaryGroupByTag(HttpServletRequest servletRequest,
       HttpServletResponse servletResponse) throws Exception {
-    return ElementsRequestExecutor.aggregateGroupByBoundaryGroupByTag(RequestResource.LENGTH,
+    return ElementsRequestExecutor.aggregateGroupByBoundaryGroupByTag(RequestResource.PERIMETER,
         servletRequest, servletResponse, true, false);
   }
 
   /**
-   * Gives the length of OSM elements grouped by the key.
+   * Gives the perimeter of polygonal OSM objects grouped by the key.
    * 
    * @param servletRequest <code>HttpServletRequest</code> of the incoming request
    * @param servletResponse <code>HttpServletResponse</code> of the outgoing response
@@ -129,22 +132,22 @@ public class LengthController {
    *         {@link org.heigit.ohsome.ohsomeapi.executor.ElementsRequestExecutor#aggregateGroupByKey(RequestResource, HttpServletRequest, HttpServletResponse, boolean, boolean)
    *         aggregateGroupByKey}
    */
-  @ApiOperation(value = "Length of OSM elements grouped by the key", nickname = "lengthGroupByKey",
-      response = GroupByResponse.class)
+  @ApiOperation(value = "Perimeter of OSM elements grouped by the key",
+      nickname = "perimeterGroupByKey", response = GroupByResponse.class)
   @ApiImplicitParams({@ApiImplicitParam(name = "groupByKeys", value = ParameterDescriptions.KEYS,
-      defaultValue = DefaultSwaggerParameters.HIGHWAY_KEY, paramType = "query", dataType = "string",
-      required = true)})
+      defaultValue = DefaultSwaggerParameters.BUILDING_KEY, paramType = "query",
+      dataType = "string", required = true)})
   @RequestMapping(value = "/groupBy/key", method = {RequestMethod.GET, RequestMethod.POST},
       produces = {"application/json", "text/csv"})
-  public Response lengthGroupByKey(HttpServletRequest servletRequest,
+  public Response perimeterGroupByKey(HttpServletRequest servletRequest,
       HttpServletResponse servletResponse) throws Exception {
-    return ElementsRequestExecutor.aggregateGroupByKey(RequestResource.LENGTH, servletRequest,
+    return ElementsRequestExecutor.aggregateGroupByKey(RequestResource.PERIMETER, servletRequest,
         servletResponse, true, false);
   }
 
   /**
-   * Gives the length of OSM elements grouped by the tag.
-   * 
+   * Gives the perimeter of polygonal OSM objects grouped by the tag.
+   *
    * @param servletRequest <code>HttpServletRequest</code> of the incoming request
    * @param servletResponse <code>HttpServletResponse</code> of the outgoing response
    * @return {@link org.heigit.ohsome.ohsomeapi.output.Response Response}
@@ -152,24 +155,24 @@ public class LengthController {
    *         {@link org.heigit.ohsome.ohsomeapi.executor.ElementsRequestExecutor#aggregateGroupByTag(RequestResource, HttpServletRequest, HttpServletResponse, boolean, boolean)
    *         aggregateGroupByTag}
    */
-  @ApiOperation(value = "Length of OSM elements grouped by the tag", nickname = "lengthGroupByTag",
-      response = GroupByResponse.class)
+  @ApiOperation(value = "Perimeter of OSM elements grouped by the tag",
+      nickname = "perimeterGroupByTag", response = GroupByResponse.class)
   @ApiImplicitParams({
       @ApiImplicitParam(name = "groupByKey", value = ParameterDescriptions.GROUP_BY_KEY,
-          defaultValue = DefaultSwaggerParameters.HIGHWAY_KEY, paramType = "query",
+          defaultValue = DefaultSwaggerParameters.BUILDING_KEY, paramType = "query",
           dataType = "string", required = true),
       @ApiImplicitParam(name = "groupByValues", value = ParameterDescriptions.VALUES,
           defaultValue = "", paramType = "query", dataType = "string", required = false)})
   @RequestMapping(value = "/groupBy/tag", method = {RequestMethod.GET, RequestMethod.POST},
       produces = {"application/json", "text/csv"})
-  public Response lengthGroupByTag(HttpServletRequest servletRequest,
+  public Response perimeterGroupByTag(HttpServletRequest servletRequest,
       HttpServletResponse servletResponse) throws Exception {
-    return ElementsRequestExecutor.aggregateGroupByTag(RequestResource.LENGTH, servletRequest,
+    return ElementsRequestExecutor.aggregateGroupByTag(RequestResource.PERIMETER, servletRequest,
         servletResponse, true, false);
   }
 
   /**
-   * Gives the density of OSM elements (length of items divided by the total area in
+   * Gives the density of OSM elements (perimeter of items divided by the total area in
    * square-kilometers).
    * 
    * @param servletRequest <code>HttpServletRequest</code> of the incoming request
@@ -179,15 +182,15 @@ public class LengthController {
    *         {@link org.heigit.ohsome.ohsomeapi.executor.AggregateRequestExecutor#aggregate() aggregate}
    */
   @ApiOperation(
-      value = "Density of OSM elements (length of elements divided by "
+      value = "Density of OSM elements (perimeter of elements divided by "
           + "the total area in square-kilometers)",
-      nickname = "lengthDensity", response = DefaultAggregationResponse.class)
+      nickname = "perimeterDensity", response = DefaultAggregationResponse.class)
   @RequestMapping(value = "/density", method = {RequestMethod.GET, RequestMethod.POST},
       produces = {"application/json", "text/csv"})
-  public Response lengthDensity(HttpServletRequest servletRequest,
+  public Response perimeterDensity(HttpServletRequest servletRequest,
       HttpServletResponse servletResponse) throws Exception {
-    AggregateRequestExecutor executor =
-        new AggregateRequestExecutor(RequestResource.LENGTH, servletRequest, servletResponse, true);
+    AggregateRequestExecutor executor = new AggregateRequestExecutor(RequestResource.PERIMETER,
+        servletRequest, servletResponse, true);
     return executor.aggregate();
   }
 
@@ -201,18 +204,22 @@ public class LengthController {
    *         {@link org.heigit.ohsome.ohsomeapi.executor.ElementsRequestExecutor#aggregateGroupByType(RequestResource, HttpServletRequest, HttpServletResponse, boolean, boolean)
    *         aggregateGroupByType}
    */
-  @ApiOperation(value = "Density of OSM elements grouped by the OSM type",
-      nickname = "lengthDensityGroupByType", response = GroupByResponse.class)
-  @RequestMapping(value = "/density/groupBy/type", method = {RequestMethod.GET, RequestMethod.POST},
+  @ApiOperation(value = "Density of OSM elements grouped by the type",
+      nickname = "perimeterDensityGroupByType", response = GroupByResponse.class)
+  @ApiImplicitParam(name = "filter", value = ParameterDescriptions.FILTER,
+      defaultValue = DefaultSwaggerParameters.BUILDING_FILTER, paramType = "query",
+      dataType = "string", required = false)
+  @RequestMapping(value = "density/groupBy/type", method = {RequestMethod.GET, RequestMethod.POST},
       produces = {"application/json", "text/csv"})
-  public Response lengthDensityGroupByType(HttpServletRequest servletRequest,
+  public Response perimeterDensityGroupByType(HttpServletRequest servletRequest,
       HttpServletResponse servletResponse) throws Exception {
-    return ElementsRequestExecutor.aggregateGroupByType(RequestResource.LENGTH, servletRequest,
+    return ElementsRequestExecutor.aggregateGroupByType(RequestResource.PERIMETER, servletRequest,
         servletResponse, true, true);
   }
 
   /**
-   * Gives density of OSM elements grouped by the boundary parameter (bounding box/circle/polygon).
+   * Gives the density of OSM elements grouped by the boundary parameter (bounding
+   * box/circle/polygon).
    * 
    * @param servletRequest <code>HttpServletRequest</code> of the incoming request
    * @param servletResponse <code>HttpServletResponse</code> of the outgoing response
@@ -222,15 +229,13 @@ public class LengthController {
    */
   @ApiOperation(
       value = "Density of OSM elements grouped by the boundary (bboxes, bcircles, or bpolys)",
-      nickname = "lengthDensityGroupByBoundary", response = GroupByResponse.class)
-  @ApiImplicitParam(name = "format", value = ParameterDescriptions.FORMAT, defaultValue = "",
-      paramType = "query", dataType = "string", required = false)
+      nickname = "perimeterDensityGroupByBoundary", response = GroupByResponse.class)
   @RequestMapping(value = "/density/groupBy/boundary",
       method = {RequestMethod.GET, RequestMethod.POST}, produces = {"application/json", "text/csv"})
-  public Response lengthDensityGroupByBoundary(HttpServletRequest servletRequest,
+  public Response perimeterDensityGroupByBoundary(HttpServletRequest servletRequest,
       HttpServletResponse servletResponse) throws Exception {
-    AggregateRequestExecutor executor =
-        new AggregateRequestExecutor(RequestResource.LENGTH, servletRequest, servletResponse, true);
+    AggregateRequestExecutor executor = new AggregateRequestExecutor(RequestResource.PERIMETER,
+        servletRequest, servletResponse, true);
     return executor.aggregateGroupByBoundary();
   }
 
@@ -244,8 +249,8 @@ public class LengthController {
    *         {@link org.heigit.ohsome.ohsomeapi.executor.ElementsRequestExecutor#aggregateGroupByBoundaryGroupByTag(RequestResource, HttpServletRequest, HttpServletResponse, boolean, boolean)
    *         aggregateGroupByBoundaryGroupByTag}
    */
-  @ApiOperation(value = "Density of OSM elements grouped by the boundary and the tag",
-      nickname = "lengthDensityGroupByBoundaryGroupByTag", response = GroupByResponse.class)
+  @ApiOperation(value = "Density of  grouped by the boundary and the tag",
+      nickname = "perimeterDensityGroupByBoundaryGroupByTag", response = GroupByResponse.class)
   @ApiImplicitParams({
       @ApiImplicitParam(name = "groupByKey", value = ParameterDescriptions.GROUP_BY_KEY,
           defaultValue = DefaultSwaggerParameters.BUILDING_KEY, paramType = "query",
@@ -254,9 +259,9 @@ public class LengthController {
           defaultValue = "", paramType = "query", dataType = "string", required = false)})
   @RequestMapping(value = "/density/groupBy/boundary/groupBy/tag",
       method = {RequestMethod.GET, RequestMethod.POST}, produces = {"application/json", "text/csv"})
-  public Response lengthDensityGroupByBoundaryGroupByTag(HttpServletRequest servletRequest,
+  public Response perimeterDensityGroupByBoundaryGroupByTag(HttpServletRequest servletRequest,
       HttpServletResponse servletResponse) throws Exception {
-    return ElementsRequestExecutor.aggregateGroupByBoundaryGroupByTag(RequestResource.LENGTH,
+    return ElementsRequestExecutor.aggregateGroupByBoundaryGroupByTag(RequestResource.PERIMETER,
         servletRequest, servletResponse, true, true);
   }
 
@@ -271,18 +276,18 @@ public class LengthController {
    *         aggregateGroupByTag}
    */
   @ApiOperation(value = "Density of OSM elements grouped by the tag",
-      nickname = "lengthDensityGroupByTag", response = GroupByResponse.class)
+      nickname = "perimeterDensityGroupByTag", response = GroupByResponse.class)
   @ApiImplicitParams({
       @ApiImplicitParam(name = "groupByKey", value = ParameterDescriptions.GROUP_BY_KEY,
-          defaultValue = DefaultSwaggerParameters.HIGHWAY_KEY, paramType = "query",
+          defaultValue = DefaultSwaggerParameters.BUILDING_KEY, paramType = "query",
           dataType = "string", required = true),
       @ApiImplicitParam(name = "groupByValues", value = ParameterDescriptions.VALUES,
           defaultValue = "", paramType = "query", dataType = "string", required = false)})
   @RequestMapping(value = "/density/groupBy/tag", method = {RequestMethod.GET, RequestMethod.POST},
       produces = {"application/json", "text/csv"})
-  public Response lengthDensityGroupByTag(HttpServletRequest servletRequest,
+  public Response perimeterDensityGroupByTag(HttpServletRequest servletRequest,
       HttpServletResponse servletResponse) throws Exception {
-    return ElementsRequestExecutor.aggregateGroupByTag(RequestResource.LENGTH, servletRequest,
+    return ElementsRequestExecutor.aggregateGroupByTag(RequestResource.PERIMETER, servletRequest,
         servletResponse, true, true);
   }
 
@@ -297,7 +302,7 @@ public class LengthController {
    *         aggregateRatio}
    */
   @ApiOperation(value = "Ratio of OSM elements satisfying filter2 within items selected by filter",
-      nickname = "lengthRatio", response = RatioResponse.class)
+      nickname = "perimeterRatio", response = RatioResponse.class)
   @ApiImplicitParams({
       @ApiImplicitParam(name = "types2", value = ParameterDescriptions.DEPRECATED_USE_FILTER2,
           defaultValue = "", paramType = "query", dataType = "string", required = false),
@@ -306,21 +311,21 @@ public class LengthController {
       @ApiImplicitParam(name = "values2", value = ParameterDescriptions.DEPRECATED_USE_FILTER2,
           defaultValue = "", paramType = "query", dataType = "string", required = false),
       @ApiImplicitParam(name = "filter", value = ParameterDescriptions.FILTER,
-          defaultValue = DefaultSwaggerParameters.HIGHWAY_FILTER2, paramType = "query",
+          defaultValue = DefaultSwaggerParameters.BUILDING_FILTER, paramType = "query",
           dataType = "string", required = false),
       @ApiImplicitParam(name = "filter2", value = ParameterDescriptions.FILTER,
-          defaultValue = DefaultSwaggerParameters.HIGHWAY_FILTER, paramType = "query",
+          defaultValue = DefaultSwaggerParameters.BUILDING_FILTER2, paramType = "query",
           dataType = "string", required = false)})
   @RequestMapping(value = "/ratio", method = {RequestMethod.GET, RequestMethod.POST},
       produces = {"application/json", "text/csv"})
-  public Response lengthRatio(HttpServletRequest servletRequest,
+  public Response perimeterRatio(HttpServletRequest servletRequest,
       HttpServletResponse servletResponse) throws Exception {
-    return ElementsRequestExecutor.aggregateRatio(RequestResource.LENGTH, servletRequest,
+    return ElementsRequestExecutor.aggregateRatio(RequestResource.PERIMETER, servletRequest,
         servletResponse);
   }
 
   /**
-   * Gives the ratio of the length of OSM elements satisfying filter2 within items selected by
+   * Gives the ratio of the perimeter of OSM elements satisfying filter2 within items selected by
    * filter grouped by the boundary.
    * 
    * @param servletRequest <code>HttpServletRequest</code> of the incoming request
@@ -330,8 +335,8 @@ public class LengthController {
    *         {@link org.heigit.ohsome.ohsomeapi.executor.ElementsRequestExecutor#aggregateRatioGroupByBoundary(RequestResource, HttpServletRequest, HttpServletResponse)
    *         aggregateRatioGroupByBoundary}
    */
-  @ApiOperation(value = "Ratio of the length of OSM elements grouped by the boundary",
-      nickname = "lengthRatioGroupByBoundary", response = RatioGroupByBoundaryResponse.class)
+  @ApiOperation(value = "Ratio of the perimeter of OSM elements grouped by the boundary",
+      nickname = "perimeterRatioGroupByBoundary", response = RatioGroupByBoundaryResponse.class)
   @ApiImplicitParams({
       @ApiImplicitParam(name = "types2", value = ParameterDescriptions.DEPRECATED_USE_FILTER2,
           defaultValue = "", paramType = "query", dataType = "string", required = false),
@@ -340,16 +345,16 @@ public class LengthController {
       @ApiImplicitParam(name = "values2", value = ParameterDescriptions.DEPRECATED_USE_FILTER2,
           defaultValue = "", paramType = "query", dataType = "string", required = false),
       @ApiImplicitParam(name = "filter", value = ParameterDescriptions.FILTER,
-          defaultValue = DefaultSwaggerParameters.HIGHWAY_FILTER2, paramType = "query",
+          defaultValue = DefaultSwaggerParameters.BUILDING_FILTER, paramType = "query",
           dataType = "string", required = false),
       @ApiImplicitParam(name = "filter2", value = ParameterDescriptions.FILTER,
-          defaultValue = DefaultSwaggerParameters.HIGHWAY_FILTER, paramType = "query",
+          defaultValue = DefaultSwaggerParameters.BUILDING_FILTER2, paramType = "query",
           dataType = "string", required = false)})
   @RequestMapping(value = "/ratio/groupBy/boundary",
       method = {RequestMethod.GET, RequestMethod.POST}, produces = {"application/json", "text/csv"})
-  public Response lengthRatioGroupByBoundary(HttpServletRequest servletRequest,
+  public Response perimeterRatioGroupByBoundary(HttpServletRequest servletRequest,
       HttpServletResponse servletResponse) throws Exception {
-    return ElementsRequestExecutor.aggregateRatioGroupByBoundary(RequestResource.LENGTH,
+    return ElementsRequestExecutor.aggregateRatioGroupByBoundary(RequestResource.PERIMETER,
         servletRequest, servletResponse);
   }
 }
