+++ src/main/lombok/org/heigit/ohsome/ohsomeapi/controller/dataextraction/features/ElementsGeometry.java
@@ -1,4 +1,4 @@
-package org.heigit.ohsome.ohsomeapi.controller.elements.features;
+package org.heigit.ohsome.ohsomeapi.controller.dataextraction.features;
 
 /** Enumeration defining the geometry of the OSM elements(RAW, BBOX, CENTROID). */
 public enum ElementsGeometry {
